create src/dojo/algorithms.
Moved te ganetic algorithm class in here.

Fix the imports.

(some files to my gitignore and some pep8)
