+++ src/axelrod_dojo/__init__.py
@@ -1,8 +1,9 @@
 from .version import __version__
 from .archetypes.fsm import FSMParams
+from .algorithms.genetic_algorithm import Population
 from .utils import (prepare_objective,
-                    Population,
                     load_params,
                     Params,
                     score_for,
                     PlayerInfo)
+
